requirements.txt
aiogram==3.5.0
asyncpg
python-dotenv
httpx
aiohttp
apscheduler
pytz
tonutils
requests
beautifulsoup4
lolzteam
rollypay